chore: restructured POM into sub-packages (admin & recruitment) and moved related pages
