+++ src/test/java/pageObjectClass/admin/AdminPage.java
@@ -1,4 +1,4 @@
-package pageObjectClass;
+package pageObjectClass.admin;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
